Update to create-dataset.sh
Without these changes, an error is thrown showing that the tif files cannot be accessed.
This is because extracting the files from the .rar file puts them in their own folder/directory inside of the data folder and the script doesn't expect that.
This change moves the files where they need to be for the script to function properly and complete without error.

Beware that it's possible this may be a case by case scenario. Perhaps on some systems the .rar files are extracted as expected while on others it is not.